pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "get-gphotos-data"
version = "0.2.0"
description = "Python desktop GUI to access Google Photos data via their API."
readme = "README.md"
requires-python = ">=3.13.11"
license = { text = "MIT" }
authors = [{ name = "Rich Lewis" }]
# Author: Rich Lewis - GitHub: @RichLewis007
dependencies = [
  "PySide6==6.7.3",
  "platformdirs>=4.2.0",
  "requests>=2.31.0",
  "google-auth>=2.23.0",
  "google-auth-oauthlib>=1.1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-qt>=4.4.0",
  "ruff>=0.5.0",
  "pyright>=1.1.380",
]

[project.scripts]
get-gphotos-data = "get_gphotos_data.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
qt_api = "pyside6"

[tool.hatch.build.targets.wheel]
packages = ["src/get_gphotos_data"]
include = ["src/get_gphotos_data/assets/**"]

[tool.hatch.build.targets.sdist]
include = [
  "src/get_gphotos_data/**",
  "tests/**",
  "docs/**",
  "CHANGELOG.md",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]
